fix: cast generate_series args to integer in nextAvailableRank
Postgres rejects parameterized generate_series($1, $2) with
"function generate_series(unknown, unknown) is not unique".
Explicit ::integer casts resolve the overload ambiguity.

Fixes #13

+++ media.js
@@ -25,7 +25,7 @@ async function nextAvailableRank(resource, userId) {
   const result = await query(
     `
     SELECT s.r AS rank
-    FROM generate_series($1, $2) AS s(r)
+    FROM generate_series($1::integer, $2::integer) AS s(r)
     WHERE NOT EXISTS (
       SELECT 1 FROM ${resource}
       WHERE user_id = $3 AND rank = s.r
